requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
Django==4.2
django-cloudinary-storage==0.3.0
django-filter==24.3
djangorestframework==3.15.2
Pillow==8.2.0
sqlparse==0.5.2
